When using a Telegraph server on an iOS device, it automatically binds the server to the device’s IP address. This means the server is accessible from outside the device (for instance, you can connect to it from other devices). This may not be desired, as it could pose a security concern.
I’ve added an option to specific the interface the server should accept connections from.
By default, I’ve set the interface to nil, which is the same as calling accept(onPort…).
This enables apps to bind the server to 127.0.0.1, so that it won’t be accessible outside the device.
If the interface is not specified, the behavior simply remains the same as before this change.